Move model logic to separate file (V2 post merge).
This required changing generateJSON to take in a prompt instead of the elements that make up a prompt and big changes to the mocking for vertex_lib_test since VertexAI is no longer a direct dependency. I also split some tests up to test just one thing each (and it made mocking easier).

Change-Id: I447cc72cb4ae503f7086e5d3df8885f70303f9a9